move mouth movement control to ChracterBehavior
