+++ avatars/avatars.js
@@ -950,12 +950,6 @@ class Avatar {
     this.lastEyeTargetTime = -Infinity;
 
     this.manuallySetMouth=false;
-    this.manuallySetMouthState=null;
-    this.manuallySetMouthStartTime=0;
-    this.manuallySetMouthAttackTime=0;
-    this.manuallySetMouthDecayTime=0;
-    this.manuallySetMouthSustainTime=0;
-    this.manuallySetMouthReleaseTime=0;
   }
   static bindAvatar(object) {
     const model = object.scene;
@@ -1891,71 +1885,6 @@ class Avatar {
       }
       this.debugMesh.visible = debug.enabled;
     }
-
-    //#################################### manually set mouth movement ##########################################
-    const _handleMouthMovementAttack=()=>{
-      this.volume = ((timestamp/1000 - this.manuallySetMouthStartTime) / this.manuallySetMouthAttackTime)/12;
-      if(timestamp/1000 - this.manuallySetMouthStartTime >= this.manuallySetMouthAttackTime){
-        this.manuallySetMouthState = 'decay';
-        this.manuallySetMouthStartTime = timestamp/1000;
-      }
-    }
-    const _handleMouthMovementDecay=()=>{
-      this.volume = (1 - ((timestamp/1000 - this.manuallySetMouthStartTime) / this.manuallySetMouthDecayTime) * 0.8)/12;
-      if(timestamp/1000 - this.manuallySetMouthStartTime >= this.manuallySetMouthDecayTime){
-        this.manuallySetMouthState='sustain';
-        this.manuallySetMouthStartTime = timestamp/1000;
-      }
-    }
-    const _handleMouthMovementSustain=()=>{
-      if(timestamp/1000 - this.manuallySetMouthStartTime >= this.manuallySetMouthSustainTime){
-        this.manuallySetMouthState='release';
-        this.manuallySetMouthStartTime = timestamp/1000;
-      } 
-    }
-    const _handleMouthMovementRelease=()=>{
-      this.volume = (0.2 - ((timestamp/1000 - this.manuallySetMouthStartTime) / this.manuallySetMouthReleaseTime) * 0.2)/12;
-      if(timestamp/1000 - this.manuallySetMouthStartTime >= this.manuallySetMouthReleaseTime){
-        this.manuallySetMouthState=null;
-        this.manuallySetMouth=false;
-        this.manuallySetMouthStartTime = -1;
-      }
-    }
-    const _handleMouthMovementNull=()=>{
-      this.manuallySetMouthState = this.manuallySetMouth ? 'attack' : null;
-      this.manuallySetMouthStartTime = timestamp/1000;
-    }
-    switch (this.manuallySetMouthState) {
-      case 'attack': {
-        _handleMouthMovementAttack();
-        break;
-      }
-      case 'decay': {
-        _handleMouthMovementDecay();
-        break;
-      }
-      case 'sustain': {
-        _handleMouthMovementSustain();
-        break;
-      }
-      case 'release': {
-        _handleMouthMovementRelease();
-        break;
-      }
-      case null: {
-        _handleMouthMovementNull();
-        break;
-      }
-    }
-
-  }
-  setMouthMoving(attack, decay, sustain, release){
-    this.manuallySetMouthState=null;
-    this.manuallySetMouth=true;
-    this.manuallySetMouthAttackTime=attack;
-    this.manuallySetMouthDecayTime=decay;
-    this.manuallySetMouthSustainTime=sustain;
-    this.manuallySetMouthReleaseTime=release;
   }
 
   isAudioEnabled() {
